+++ tests/Regressors/ExtraTreeRegressorTest.php
@@ -128,7 +128,7 @@ class ExtraTreeRegressorTest extends TestCase
     public function params() : void
     {
         $expected = [
-            'max_depth' => 10,
+            'max_height' => 10,
             'max_leaf_size' => 3,
             'max_features' => 6,
             'min_purity_increase' => 1.0E-7,